valleyisland-io: fix SD card stuck in runtime suspend
A host controller for a SD card may need a GPIO
for card detect in order to wake up from runtime
suspend when a card is inserted.  If that GPIO is
not configured, then the host controller will not
wake up.  Fix that for the affected devices by not
enabling runtime PM unless the GPIO is successfully
set up.

Signed-off-by: Ong Boon Leong <boon.leong.ong@intel.com>