configs: drop CONFIG_MII
CONFIG_MII is a tristate with no user prompt, as such, it is
only selected by other kernel configs. If we have it in our fragments
we'll be warned when it isn't in the final .config.

The warning isn't useful, since if this is needed another option will
select it.

To avoid the warning, we just drop CONFIG_MII from our fragments

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>