qemuppc: drop CONFIG_PNP
CONFIG_PNP will be excluded from the final .config via our
lack of ISA support. We don't need this feature (since it
is disabled), and due to the way Kconfig processes options
CONFIG_PNP is not it the final .config at all.

This lack of "is not set" triggers a warning. To avoid the
warning, we drop the option.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>