bsp/common-pc: Enable R8169 driver since its needed by supported platforms
The Realtek 8169 driver is needed by the Minnowboard MAX, one of the
boards we currently support, since some of our images do not contain
modules by default, the network isnt working on genericx86, genericx86-64
and intel-core2-32 (meta-intel) when no modules are installed.

This patch fixes network on images not containing modules when
using the previously mentioned MACHINES for this board.

Signed-off-by: Alejandro Hernandez <alejandro.hernandez@linux.intel.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>