features: Fix configcheck warnings in features used by intel-quark BSPs
In mei-me.cfg, CONFIG_PCI was mistakenly set to =m when its a boolean.

In bosch-pressure-sensor-i2c, change CONFIG_BMP085 to =m instead of =y,
as its dependency is a module. Also attempt to disable BMP280, as it
conflicts with the BMP085 driver and throws a warning.

In x1000.cfg, remove CONSOLE related 'is not set' CONFIGs. VGA_CONSOLE
can't be disabled without enabling EXPERT (and therefore DEBUG_KERNEL),
which we don't want normally. The other two configurations are missing
dependencies, so they don't get set anyway, but cause warnings.

Signed-off-by: California Sullivan <california.l.sullivan@intel.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>