leds.cfg: Remove CONFIG_LEDS_TRIGGER_CPU
This is incompatible with preempt-rt and would be a mess to fix any
other way. If we were to move CONFIG_LEDS_TRIGGER_CPU to a new feature
such as leds-full, it would require us to add leds-full to every non-rt
BSP wanting leds support, and -rt BSPs would need the feature without
it. Setting "# CONFIG_LEDS_TRIGGER_CPU is not set" also doesn't work, as
it depends on !PREEMPT_RT_BASE, making it completely unavailable for
setting and still giving us a warning in kernel_configcheck.

All this option does is allow CPUs to trigger LEDs based on their usage,
so it is not a major loss anyway.

Signed-off-by: California Sullivan <california.l.sullivan@intel.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>