netfilter: enable connection tracking for IPv6
In case of the DROP policy in the INPUT chain a host using IPv6 still
might need to receive TCP packets for established connections, that is
to have the rule

-A INPUT -m conntrack --ctstate RELATED,ESTABLISHED -j ACCEPT

in its INPUT chain of ip6tables. For this feature to work the option
CONFIG_NF_CONNTRACK_IPV6 needs to be enabled.

Signed-off-by: Dmitry Rozhkov <dmitry.rozhkov@linux.intel.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>