intel-quark: Add intel-quark-preempt-rt bsp configuration
While there are intel-quark configurations for the KTYPEs standard and
tiny in bsp/intel-common, there's none for the preempt-rt KTYPE.
Trying to build preempt-rt enabled kernels such as linux-yocto-rt for
intel-quark yields a .config having a potentially misconfigured
architecture. More importantly, however, preempt-rt related CONFIG
options are not enabled. Hence, a build of, e.g., linux-yocto-rt, does
not result in a preempt-rt enabled kernel.

This patch qualifies to be (back)ported to other branches than master.

Signed-off-by: Christian Storm <christian.storm@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>