smp: Separate smp into 32 and 64 bit versions to avoid kernel warnings
When selecting SMP, some dependencies are needed, in the case
of a 32 bit build CONFIG_X86_BIGSMP=y is required, but adding that
to a 64 bit build causes warnings to be shown since X86_BIGSMP
requires X86_32 [n] itself, which can only be true for a 32 bit build.

This patch creates a 64 bit version of smp (smp_64), avoiding the
dependecy issues.

[YOCTO #11742]

Signed-off-by: Alejandro Hernandez <alejandro.hernandez@linux.intel.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>