bsp/xilinx/zynq: Add a general purpose KMACHINE for Zynq targets
Add a general purpose KMACHINE that can be used for general purpose Zynq
BSPs. The main configuration of the Zynq SoC can be configured via
device tree, allowing for a common kernel config. This allows for a
single KMACHINE to target multiple BSPs without needing to be directly
configured for the platform/soc.

This adds configs for -standard and -tiny.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>