features/debug/debug-dyndbg: Add feature to enable dynamic debug
Add a feature to enable the kernel DYNAMIC_DEBUG config which allows for
dynamic printing of debug messages. Kernel debug messages are normally
not enabled in the kernel unless manually defining DEBUG in code files.
Dynamic debug allows for the messages to be embedded but toggled
dynamically during run time (via boot parameters or from a sysfs
interface).

Enabling this option is useful when debugging multiple parts of the
kernel as well as to allow for enabling of debug messages without
needing to modify the source.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>