qemuarma9: drop VGA console disable
The Kconfig's for VGA console have changed such that they can't
be set (or unset) for architectures/platforms that have no support.

As such, we don't need to unset it anymore.

Having the unset triggers a warning during kernel processing, so
the choice is easy .. remove it from the fragment.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>