features/i915/i915.cfg: compile i915 as a module
Set i915 as a module to trigger the firmware load
at the same time the module is loaded.

This can avoid a timing problem between the driver starting
and it triggering a firmware load, after compile it as module
i915 driver will start after the rootfs is ready, then the
firmware store in /lib/firmware/ can be load successful by
i915 driver.

Signed-off-by: Liwei Song <liwei.song@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>