mpc8315e-rdb: Enable Realtek PHY driver
A Realtek RTL8211B PHY is integrated on mpc8315erdb board. It works
pretty well with the genphy driver. But now we apply a patch for the
Realtek PHY driver in order to fix an Ethernet broken issue on this
board. So enable this chip specific driver to make sure that fix can
take effect.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>