wifi: move patches to a dedicated "patches" .scc file
With the reorganization of the wifi configuration fragments, it is common
that wifi is enabled via KERNEL_FEATURES inclusion. This technique works
on 4.14/4.15/master, but not on 4.12, since the iwlwifi .scc file contains
patches .. which are already applied to the tree, as a result you get
patch failures when attempting to add the wifi configuratio to your kernel.

To stay consistent with 4.14+, we can move the patches to a iwlwifi-patches
file, include that from the BSPs (so patches will be applied during a new
kernel tree construction). We then leave iwlwifi as just configuration
fragments (which are also included by the BSPs), and that configuration
fragment is safe to use via KERNEL_FEATURES.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>