arm64: dts: Fix broken architected timer interrupt trigger
Issue: LIN10-3946

1/1 [
Author: Zumeng Chen
Email: zumeng.chen@windriver.com
Subject: arm64: dts: Fix broken architected timer interrupt trigger
Date: Tue, 22 May 2018 16:08:49 +0800

commit f2a89d3b2b85b90b05453872aaabfdb412a21a03 upstream

The ARM architected timer specification mandates that the interrupt
associated with each timer is level triggered (which corresponds to
the "counter >= comparator" condition).

A number of DTs are being remarkably creative, declaring the interrupt
to be edge triggered. A quick look at the TRM for the corresponding ARM
CPUs clearly shows that this is wrong, and I've corrected those.
For non-ARM designs (and in the absence of a publicly available TRM),
I've made them active low as well, which can't be completely wrong
as the GIC cannot disinguish between level low and level high.

The respective maintainers are of course welcome to prove me wrong.

While I was at it, I took the liberty to fix a couple of related issue,
such as some spurious affinity bits on ThunderX, and their complete
absence on ls1043a (both of which seem to be related to copy-pasting
from other DTs).

Acked-by: Duc Dang <dhdang@apm.com>
Acked-by: Carlo Caione <carlo@endlessm.com>
Acked-by: Michal Simek <michal.simek@xilinx.com>
Acked-by: Krzysztof Kozlowski <k.kozlowski@samsung.com>
Acked-by: Dinh Nguyen <dinguyen@opensource.altera.com>
Acked-by: Masahiro Yamada <yamada.masahiro@socionext.com>
Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>
Signed-off-by: Arnd Bergmann <arnd@arndb.de>
[Partly for zynqmp]
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>