MIPS: Use '+=" instead of '=' to avoid the CFLAGS override
1/1 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: MIPS: Use '+=" instead of '=' to avoid the CFLAGS override
Date: Mon, 11 Jun 2018 18:17:54 +0800

We used the CFLAGS_xxx to workaround the gcc 8 build warnings
for some specific file. But CFLAGS_xxx is also used with '=' in
other places of this Makefile. This override the gcc 8 workaround,
so replace all the '=' with '+=" to fix this issue.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>