ext4: only look at the bg_flags field if it is valid
Issue: LIN10-4429

1/2 [
Author: Theodore Ts'o
Email: tytso@mit.edu
Subject: ext4: don't update checksum of new initialized bitmaps
Date: Mon, 19 Feb 2018 14:16:47 -0500

commit 044e6e3d74a3d7103a0c8a9305dfd94d64000660 upstream

When reading the inode or block allocation bitmap, if the bitmap needs
to be initialized, do not update the checksum in the block group
descriptor.  That's because we're not set up to journal those changes.
Instead, just set the verified bit on the bitmap block, so that it's
not necessary to validate the checksum.

When a block or inode allocation actually happens, at that point the
checksum will be calculated, and update of the bg descriptor block
will be properly journalled.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Cc: stable@vger.kernel.org
Signed-off-by: Rui Wang <rui.wang@windriver.com>
]

2/2 [
Author: Theodore Ts'o
Email: tytso@mit.edu
Subject: ext4: only look at the bg_flags field if it is valid
Date: Thu, 14 Jun 2018 00:58:00 -0400

commit 8844618d8aa7a9973e7b527d038a2a589665002c upstream

The bg_flags field in the block group descripts is only valid if the
uninit_bg or metadata_csum feature is enabled.  We were not
consistently looking at this field; fix this.

Also block group #0 must never have uninitialized allocation bitmaps,
or need to be zeroed, since that's where the root inode, and other
special inodes are set up.  Check for these conditions and mark the
file system as corrupted if they are detected.

This addresses CVE-2018-10876.

https://bugzilla.kernel.org/show_bug.cgi?id=199403

Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Cc: stable@kernel.org
Signed-off-by: Rui Wang <rui.wang@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>