wifi_ad_hoc will cause board calltrace on Whiskey Lake
Issue: LIN10-4557

1/4 [
Author: Sara Sharon
Email: sara.sharon@intel.com
Subject: iwlwifi: mvm: fix IBSS for devices that support station type API
Date: Thu, 21 Dec 2017 15:05:28 +0200

commit fc07bd8ce19bff9e7479c04077ddb5957d1a27be upstream.

In IBSS, the mac80211 sets the cab_queue to be invalid.

However, the multicast station uses it, so we need to override it.

A previous patch did it, but it was nested inside the if's and was
applied only for legacy FWs that don't support the new station type
API, instead of being applied for all paths.

In addition, add a missing NL80211_IFTYPE_ADHOC to the initialization
of the queues in iwl_mvm_mac_ctxt_init()

Fixes: ee48b72211f8 ("iwlwifi: mvm: support ibss in dqa mode")
Signed-off-by: Sara Sharon <sara.sharon@intel.com>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Signed-off-by: Liwei Song <liwei.song@windriver.com>
]

2/4 [
Author: Gregory Greenman
Email: gregory.greenman@intel.com
Subject: iwlwifi: mvm: rs: introduce new API for rate scaling
Date: Wed, 1 Nov 2017 07:16:29 +0200

commit ecaf71de414345c68a05c403d33dd0cd89b2c85f upstream.

New devices will have rate scaling algorithm running in the firmware.
With this feature, the driver's responsiblity is to provide an initial
configuration and to handle notifications regarding recent rates and
some other parameters. Debugfs hooks will be still available for
reading the current rate/statistics and setting a fixed rate.
The old API is supported so far, though both APIs cannot be used
simultaneously.

This is the first patch in the series. It adds a new TLV specifying
FW support for the new API and updates lq_sta to support two types
of rate scaling.

Signed-off-by: Gregory Greenman <gregory.greenman@intel.com>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Signed-off-by: Liwei Song <liwei.song@windriver.com>
]

3/4 [
Author: Shaul Triebitz
Email: shaul.triebitz@intel.com
Subject: iwlwifi: mvm: adjust to quota offload
Date: Mon, 4 Dec 2017 15:33:46 +0200

commit dad3340fd988173aa79afb03941416c85f82ac0c upstream.

Skip Quota handling if firmware implements Quota.

Signed-off-by: Shaul Triebitz <shaul.triebitz@intel.com>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Signed-off-by: Liwei Song <liwei.song@windriver.com>
]

4/4 [
Author: Emmanuel Grumbach
Email: emmanuel.grumbach@intel.com
Subject: iwlwifi: bump the max API version for 9000 and 22000 devices
Date: Mon, 5 Feb 2018 12:44:44 +0200

commit 8f27036a0e380c6ede67a8d2a132327fcb760ace upstream.

We are now ready to load 38.ucode

Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
[Adjust patch for file name(a000) changed]
Signed-off-by: Liwei Song <liwei.song@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>