tipc: fix the big/little endian issue in tipc_dest
Issue: LIN10-4746

1/1 [
Author: Haiqing Bai
Email: Haiqing.Bai@windriver.com
Subject: tipc: fix the big/little endian issue in tipc_dest
Date: Mon, 27 Aug 2018 09:32:26 +0800

commit 30935198b7d0be12b1c45c328b66a7fdefb16256 upstream

In function tipc_dest_push, the 32bit variables 'node' and 'port'
are stored separately in uppper and lower part of 64bit 'value'.
Then this value is assigned to dst->value which is a union like:
union
{
  struct {
    u32 port;
    u32 node;
  };
  u64 value;
}
This works on little-endian machines like x86 but fails on big-endian
machines.

The fix remove the 'value' stack parameter and even the 'value'
member of the union in tipc_dest, assign the 'node' and 'port' member
directly with the input parameter to avoid the endian issue.

Fixes: a80ae5306a73 ("tipc: improve destination linked list")
Signed-off-by: Zhenbo Gao <zhenbo.gao@windriver.com>
Acked-by: Jon Maloy <jon.maloy@ericsson.com>
Signed-off-by: Haiqing Bai <Haiqing.Bai@windriver.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>