edac: release some memories which registered or not to the driver subsystem.
Issue: LIN1018-3070

1/1 [
Author: Jun Miao
Email: jun.miao@windriver.com
Subject: edac: release some memories which registered or not to the driver subsystem.
Date: Mon, 17 Dec 2018 22:10:13 -0500

This patch can fix the following kmemleak:

unreferenced object 0xffffa1f639d645e0 (size 32):
  comm "systemd-udevd", pid 361, jiffies 4294679877 (age 653.289s)
  hex dump (first 32 bytes):
    00 00 00 00 00 00 00 00 00 c0 30 32 f6 a1 ff ff ..........02....
    00 c8 30 32 f6 a1 ff ff 00 00 00 00 00 00 00 00 ..02............
  backtrace:
    [<0000000085cb7a6c>] kmem_cache_alloc_trace+0x148/0x200
    [<00000000c3e7e24d>] edac_mc_alloc+0x316/0x660
    [<00000000a785dc39>] 0xffffffffc03b7566
    [<00000000e2a8abd5>] do_one_initcall+0x53/0x1fe
    [<00000000cc6e09bd>] do_init_module+0x60/0x210
    [<0000000094a5f135>] load_module+0x215b/0x2650
    [<00000000da96f61b>] __se_sys_finit_module+0xc3/0xe0
    [<0000000004a9f952>] __x64_sys_finit_module+0x1a/0x20
    [<00000000e2af0e30>] do_syscall_64+0x4d/0xf0
    [<00000000dee0ce8b>] entry_SYSCALL_64_after_hwframe+0x44/0xa9
    [<00000000745c7f7a>] 0xffffffffffffffff

we have observed some memory leaks when removing the skx_edac.ko.
Actually there are two kinds of memory leaks. The first one is that
some memories used by the csrow and dimm devices are not released in
the corresponding release callback function when the device is unregistered
by using device_unregister(). The second is that some csrow and dimm
devices are not registered to the device driver subsystem at all in some
special cases, so we have to release all the memories used by csrow and
dimm devices.

Signed-off-by: Jun Miao <jun.miao@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>