check console device file on fs when booting
1/1 [
Author: Bruce Ashfield
Email: bruce.ashfield@windriver.com
Subject: check console device file on fs when booting
Date: Thu, 8 Apr 2010 23:44:21 -0700

If a root filesystem is generated as non-root, one of
the tell tale signs is /dev/console not being a character
file. To save a whole class of questions, let's just test
for the condition and let the user know.

Signed-off-by: Richard Laroque <rlarocqu@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>
]
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>