ARC: Add nsimhs- and hsdk-standard configs
With ARC support ramping-up in upstream OE we're ready to
build more complicated distributions and linux-yocto is
a nice and configurable base for that.

This commit adds support of 1 simulated board (nSIM) and
the mast affordable and powerful to date development board (HSDK).

Once Qemu port for ARC is functional (it's being actively developed now)
we'll switch from nSIM to Qemu.

Still it would be really good to keep nSIM support in linux-yocto for now
as it allows for simpler testing as compared to real HW.

Signed-off-by: Alexey Brodkin <abrodkin@synopsys.com>
Cc: Bruce Ashfield <bruce.ashfield@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>