NVDIMM: Label read iterations can be reduced by requesting larger chunks
Issue: LIN1018-4261

1/8 [
Author: Alexander Duyck
Email: alexander.h.duyck@linux.intel.com
Subject: libnvdimm: Set device node in nd_device_register
Date: Tue, 25 Sep 2018 13:53:07 -0700

commit 1a091d16dbff6d373c5335e08316756e6c39f048 upstream.

This change makes it so that we don't repeatedly overwrite the device node
for nvdimm regions. The earliest we can set the node is immediately after
calling device init, so I have moved the code there so we can avoid
rewriting the node with each uevent.

Signed-off-by: Alexander Duyck <alexander.h.duyck@linux.intel.com>
Signed-off-by: Dan Williams <dan.j.williams@intel.com>
Signed-off-by: Yongxin Liu <yongxin.liu@windriver.com>
]

2/8 [
Author: Alexander Duyck
Email: alexander.h.duyck@linux.intel.com
Subject: nvdimm: Sanity check labeloff
Date: Wed, 10 Oct 2018 16:38:41 -0700

commit d86d4d63d88861107d3bfc84be7294552231ecd0 upstream.

This patch adds validation for the labeloff field in the indexes.

Reviewed-by: Toshi Kani <toshi.kani@hpe.com>
Signed-off-by: Alexander Duyck <alexander.h.duyck@linux.intel.com>
Signed-off-by: Dan Williams <dan.j.williams@intel.com>
Signed-off-by: Yongxin Liu <yongxin.liu@windriver.com>
]

3/8 [
Author: Alexander Duyck
Email: alexander.h.duyck@linux.intel.com
Subject: nvdimm: Clarify comment in sizeof_namespace_index
Date: Wed, 10 Oct 2018 16:38:55 -0700

commit 1cfeb66e8e137be8e01b88bb4d416e987abda4a4 upstream.

When working on the label code I found it rather confusing to see several
spots that reference a minimum label size of 256 while working with labels
that are 128 bytes in size.

This patch is meant to provide a clarification on one of the comments that
was at the heart of the issue. Specifically for version 1.2 and later of
the namespace specification the minimum label size is 256, prior to that
the minimum label size was 128. So we should state that as such to avoid
confusion.

Reviewed-by: Toshi Kani <toshi.kani@hpe.com>
Signed-off-by: Alexander Duyck <alexander.h.duyck@linux.intel.com>
Signed-off-by: Dan Williams <dan.j.williams@intel.com>
Signed-off-by: Yongxin Liu <yongxin.liu@windriver.com>
]

4/8 [
Author: Alexander Duyck
Email: alexander.h.duyck@linux.intel.com
Subject: nvdimm: Remove empty if statement
Date: Wed, 10 Oct 2018 16:39:06 -0700

commit 19418b024427ec60ba6084addf691a8d93670398 upstream.

This patch removes an empty statement from an if expression and promotes
the else statement to the if expression with the expression logic reversed.

I feel this is more readable as the empty statement can lead to issues if
any additional logic was ever added.

Reviewed-by: Toshi Kani <toshi.kani@hpe.com>
Signed-off-by: Alexander Duyck <alexander.h.duyck@linux.intel.com>
Signed-off-by: Dan Williams <dan.j.williams@intel.com>
Signed-off-by: Yongxin Liu <yongxin.liu@windriver.com>
]

5/8 [
Author: Dan Williams
Email: dan.j.williams@intel.com
Subject: libnvdimm: Introduce locked DIMM capacity support
Date: Wed, 13 Jun 2018 09:08:36 -0700

commit 08e6b3c6e3a054f566367740c94b8c1d18e52056 upstream.

When a DIMM is locked its namespace label area may not be. Introduce the
distinction of locked namespaces to allow namespace enumeration while
the capacity is locked.

Signed-off-by: Dan Williams <dan.j.williams@intel.com>
Signed-off-by: Yongxin Liu <yongxin.liu@windriver.com>
]

6/8 [
Author: Dan Williams
Email: dan.j.williams@intel.com
Subject: libnvdimm, dimm: Maximize label transfer size
Date: Wed, 10 Oct 2018 16:38:24 -0700

commit d11cf4a7321b538563b0ab30dc0d1f18f9c56226 upstream.

Use kvzalloc() to bypass the arbitrary PAGE_SIZE limit of label transfer
operations. Given the expense of calling into firmware, maximize the
amount of label data we transfer per call to be up to the total label
space if allowed by the firmware.

Instead of limiting based on PAGE_SIZE we can instead simply limit the
maximum size based on either the config_size int he case of the get
operation, or the length of the write based on the set operation.

On a system with 24 NVDIMM modules each with a config_size of 128K and a
maximum transfer size of 64K - 4, this patch reduces the init time for the
label data from around 24 seconds down to between 4-5 seconds.

Reviewed-by: Toshi Kani <toshi.kani@hpe.com>
Signed-off-by: Alexander Duyck <alexander.h.duyck@linux.intel.com>
Signed-off-by: Dan Williams <dan.j.williams@intel.com>
Signed-off-by: Yongxin Liu <yongxin.liu@windriver.com>
]

7/8 [
Author: Alexander Duyck
Email: alexander.h.duyck@linux.intel.com
Subject: nvdimm: Split label init out from the logic for getting config data
Date: Wed, 10 Oct 2018 16:39:20 -0700

commit 2d657d17f72d2ae70c02f0d0ea6a04ad0f016b57 upstream.

This patch splits the initialization of the label data into two functions.
One for doing the init, and another for reading the actual configuration
data. The idea behind this is that by doing this we create a symmetry
between the getting and setting of config data in that we have a function
for both. In addition it will make it easier for us to identify the bits
that are related to init versus the pieces that are a wrapper for reading
data from the ACPI interface.

So for example by splitting things out like this it becomes much more
obvious that we were performing checks that weren't necessarily related to
the set/get operations such as relying on ndd->data being present when the
set and get ops should not care about a locally cached copy of the label
area.

Reviewed-by: Toshi Kani <toshi.kani@hpe.com>
Signed-off-by: Alexander Duyck <alexander.h.duyck@linux.intel.com>
Signed-off-by: Dan Williams <dan.j.williams@intel.com>
Signed-off-by: Yongxin Liu <yongxin.liu@windriver.com>
]

8/8 [
Author: Alexander Duyck
Email: alexander.h.duyck@linux.intel.com
Subject: nvdimm: Use namespace index data to reduce number of label reads needed
Date: Wed, 10 Oct 2018 16:39:35 -0700

commit 7d47aad4570e5e6e9a8162bb417ca9b74132f27c upstream.

This patch adds logic that is meant to make use of the namespace index data
to reduce the number of reads that are needed to initialize a given
namespace. The general idea is that once we have enough data to validate
the namespace index we do so and then proceed to fetch only those labels
that are not listed as being "free". By doing this I am seeing a total time
reduction from about 4-5 seconds to 2-3 seconds for 24 NVDIMM modules each
with 128K of label config area.

Reviewed-by: Toshi Kani <toshi.kani@hpe.com>
Signed-off-by: Alexander Duyck <alexander.h.duyck@linux.intel.com>
Signed-off-by: Dan Williams <dan.j.williams@intel.com>
Signed-off-by: Yongxin Liu <yongxin.liu@windriver.com>
]
Signed-off-by: Yue Tao <Yue.Tao@windriver.com>