bsp/beaglebone: support qemu -machine virt
Purpose: Support booting a beaglebone image with runqemu.

If the kernel supports Qemu's virt machine, runqemu works almost for free.
The device tree for machine virt is included in Qemu, which simplifies
everything quite a bit.
This change adds ARCH_VIRT=y and some drivers to the beaglebone kernel
configuration which allows to:

  export MACHINE="beaglebone-yocto"
  bitbake core-image-minimale
  runqemu

This also works out of an eSDK. Whithout this feature usually two different
SDKs need to be compiled and maintained. One SDK is used for development
in Qemu, another one is used to develop for the real target hardware.

[Yocto #13384]

Signed-off-by: Adrian Freihofer <adrian.freihofer@siemens.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>