net: mvpp2: Restore the BM pool id of the rxq overrode by the MUSDK
Issue: LIN1018-4260

1/1 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: net: mvpp2: Restore the BM pool id of the rxq overrode by the MUSDK
Date: Wed, 19 Jun 2019 09:12:29 +0800

When a port is hijacked by the MUSDK, the BM pool id of the rxq will
be set to the one created by the MUSDK, so we should restore the BM
pool id when switching back to kernel. Otherwise the kernel would panic
due to the buffs from MUSDK are used.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]
Signed-off-by: Yue Tao <Yue.Tao@windriver.com>