floppy: fix out-of-bounds read in copy_buffer
Issue: LIN1018-4538

1/1 [
Author: Denis Efremov
Email: efremov@ispras.ru
Subject: floppy: fix out-of-bounds read in copy_buffer
Date: Fri, 12 Jul 2019 21:55:23 +0300

commit da99466ac243f15fbba65bd261bfc75ffa1532b6 upstream

This fixes a global out-of-bounds read access in the copy_buffer
function of the floppy driver.

The FDDEFPRM ioctl allows one to set the geometry of a disk.  The sect
and head fields (unsigned int) of the floppy_drive structure are used to
compute the max_sector (int) in the make_raw_rw_request function.  It is
possible to overflow the max_sector.  Next, max_sector is passed to the
copy_buffer function and used in one of the memcpy calls.

An unprivileged user could trigger the bug if the device is accessible,
but requires a floppy disk to be inserted.

The patch adds the check for the .sect * .head multiplication for not
overflowing in the set_geometry function.

The bug was found by syzkaller.

Signed-off-by: Denis Efremov <efremov@ispras.ru>
Tested-by: Willy Tarreau <w@1wt.eu>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Ovidiu Panait <ovidiu.panait@windriver.com>
]
Signed-off-by: Yue Tao <Yue.Tao@windriver.com>