block: blk_init_allocated_queue() set q->fq as NULL in the fail case
Issue: LIN10-6198

1/1 [
Author: xiao jin
Email: jin.xiao@intel.com
Subject: block: blk_init_allocated_queue() set q->fq as NULL in the fail case
Date: Mon, 30 Jul 2018 14:11:12 +0800

commit 54648cf1ec2d7f4b6a71767799c45676a138ca24 upstream

We find the memory use-after-free issue in __blk_drain_queue()
on the kernel 4.14. After read the latest kernel 4.18-rc6 we
think it has the same problem.

Memory is allocated for q->fq in the blk_init_allocated_queue().
If the elevator init function called with error return, it will
run into the fail case to free the q->fq.

Then the __blk_drain_queue() uses the same memory after the free
of the q->fq, it will lead to the unpredictable event.

The patch is to set q->fq as NULL in the fail case of
blk_init_allocated_queue().

Fixes: commit 7c94e1c157a2 ("block: introduce blk_flush_queue to drive flush machinery")
Cc: <stable@vger.kernel.org>
Reviewed-by: Ming Lei <ming.lei@redhat.com>
Reviewed-by: Bart Van Assche <bart.vanassche@wdc.com>
Signed-off-by: xiao jin <jin.xiao@intel.com>
Signed-off-by: Jens Axboe <axboe@kernel.dk>
Signed-off-by: Zhixiong Chi <zhixiong.chi@windriver.com>
]
Signed-off-by: Yue Tao <Yue.Tao@windriver.com>