drivers/virt/fsl_hypervisor.c: prevent integer overflow in ioctl
Issue: LIN10-6220

1/1 [
Author: Dan Carpenter
Email: dan.carpenter@oracle.com
Subject: drivers/virt/fsl_hypervisor.c: prevent integer overflow in ioctl
Date: Tue, 14 May 2019 15:47:03 -0700

commit 6a024330650e24556b8a18cc654ad00cfecf6c6c upstream

The "param.count" value is a u64 thatcomes from the user.  The code
later in the function assumes that param.count is at least one and if
it's not then it leads to an Oops when we dereference the ZERO_SIZE_PTR.

Also the addition can have an integer overflow which would lead us to
allocate a smaller "pages" array than required.  I can't immediately
tell what the possible run times implications are, but it's safest to
prevent the overflow.

Link: http://lkml.kernel.org/r/20181218082129.GE32567@kadam
Fixes: 6db7199407ca ("drivers/virt: introduce Freescale hypervisor management driver")
Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>
Reviewed-by: Andrew Morton <akpm@linux-foundation.org>
Cc: Timur Tabi <timur@freescale.com>
Cc: Mihai Caraman <mihai.caraman@freescale.com>
Cc: Kumar Gala <galak@kernel.crashing.org>
Cc: <stable@vger.kernel.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Wenlin Kang <wenlin.kang@windriver.com>
]
Signed-off-by: Yue Tao <Yue.Tao@windriver.com>